requirements.txt
requests~=2.24.0
python-dateutil~=2.8.1
setuptools~=50.3.2
boto3~=1.18.4